+++ std/examples/tests/xeval_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 import { xeval } from "../xeval.ts";
 import { stringsReader } from "../../io/util.ts";
-import { decode, encode } from "../../strings/mod.ts";
+import { decode, encode } from "../../encoding/utf8.ts";
 import {
   assertEquals,
   assertStrContains,
