+++ std/fs/expand_glob_test.ts
@@ -1,5 +1,5 @@
 const { cwd, execPath, run } = Deno;
-import { decode } from "../strings/mod.ts";
+import { decode } from "../encoding/utf8.ts";
 import { assert, assertEquals, assertStrContains } from "../testing/asserts.ts";
 import {
   isWindows,
