+++ std/io/readers_test.ts
@@ -3,7 +3,7 @@ import { assertEquals } from "../testing/asserts.ts";
 import { MultiReader, StringReader } from "./readers.ts";
 import { StringWriter } from "./writers.ts";
 import { copyN } from "./ioutil.ts";
-import { decode } from "../strings/mod.ts";
+import { decode } from "../encoding/utf8.ts";
 
 test(async function ioStringReader(): Promise<void> {
   const r = new StringReader("abcdef");
