--- std/strings/decode.ts
@@ -1,7 +1,0 @@
-/** A default TextDecoder instance */
-export const decoder = new TextDecoder();
-
-/** Shorthand for new TextDecoder().decode() */
-export function decode(input?: Uint8Array): string {
-  return decoder.decode(input);
-}
