--- std/strings/mod.ts
@@ -1,4 +1,0 @@
-// Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-
-export * from "./encode.ts";
-export * from "./decode.ts";
